requirements.txt
annotated-types==0.7.0
anthropic==0.40.0
anyio==4.7.0
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
distro==1.9.0
filelock==3.16.1
fsspec==2024.10.0
googletrans==4.0.0rc1
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.12.1
httpcore==1.0.7
httpx==0.28.1
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.4
jiter==0.8.2
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.57.4
openai-whisper==20240930
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
sniffio==1.3.1
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.2.3
yt-dlp==2024.12.6
